% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_credentials_to_token.R
\name{api_credentials_to_token}
\alias{api_credentials_to_token}
\title{Create API token.}
\usage{
api_credentials_to_token(
  app_name,
  consumer_key,
  consumer_secret,
  access_token,
  access_secret
)
}
\arguments{
\item{app_name}{Twitter app name.}

\item{consumer_key}{Twitter consumer key.}

\item{consumer_secret}{Twitter consumer secret.}

\item{access_token}{Twitter access token.}

\item{access_secret}{Twitter access secret.}
}
\value{
Twitter API token.
}
\description{
api_crendentials_to_token takes Twitter API credentials and returns an API token. This has to be entered into the token argument when running the get_user_network() function.
}
\examples{
\dontrun{

ex_app_name = "application_name"
ex_consumer_key = ####
ex_consumer_secret = ####
ex_access_token = ####
ex_access_secret = ####

api_credentials_to_token(app_name = ex_app_name, consumer_key = ex_consumer_key,
                         consumer_secret = ex_consumer_secret, access_token = ex_access_token,
                         access_secret = ex_access_secret)
}
}
