% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_network.R
\name{get_user_network}
\alias{get_user_network}
\title{Takes Twitter user information and returns list of data on users in network connections and an edgelist of user's network.}
\usage{
get_user_network(
  id = NULL,
  degree = 1,
  token = NULL,
  track_progress = TRUE,
  filter_col = NULL,
  filter_val = NULL,
  greater = TRUE,
  base_nodes_edges = NULL,
  layer_count = 0,
  collected_ids = NULL,
  steps = 1
)
}
\arguments{
\item{id}{Twitter user id.}

\item{degree}{Number of "layers" of user's friends/followers to collect.}

\item{token}{Twitter API token.}

\item{track_progress}{Determines whether to print progress of data collection (optional).}

\item{filter_col}{Name of column of user tibble to filter for Twitter user collection (optional).}
}
\value{
List with two elements: user tibble and edgelist tibble.
}
\description{
Takes Twitter user information and returns list of data on users in network connections and an edgelist of user's network.
}
\examples{
ex_id <- 778619636510326784 # Example Twitter user id
ex_degree <- 2 # Collect user's friends and their friends' friends
ex_token <- api_token # Returned from api_credentials_to_token function
get_user_network(id = ex_id, degree = ex_degree, token = ex_token)
}
