% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_network.R
\name{get_user_network}
\alias{get_user_network}
\title{Get Twitter user network data}
\usage{
get_user_network(
  screen_name = NULL,
  id = NULL,
  degree = 1,
  token = NULL,
  filter_col = NULL,
  filter_val = NULL,
  filter_logic = "any",
  greater = TRUE,
  ...
)
}
\arguments{
\item{screen_name}{Twitter user screen name.}

\item{id}{Twitter user id.}

\item{degree}{Number of degrees of user's friends/followers to collect.}

\item{token}{Twitter API token.}

\item{filter_col}{Name of column of user tibble to filter for Twitter user collection (optional).}

\item{filter_val}{Value to filter on (optional).}

\item{filter_logic}{For multiple filters, determines whether to filter based on "any" or "all" criteria (optional).}

\item{greater}{When filtering number of friends/followers, determines whether filter is ceiling or floor on number of users (optional).}

\item{...}{Additional arguments used for recursive function call within function.}
}
\value{
List with two elements: user tibble and edgelist tibble.
}
\description{
get_user_network() takes a Twitter profile name or id number and returns data on this user's network connections, including an edgelist of the user's network.
}
\examples{
\dontrun{

# Collect all connections 2 degrees from Twitter user Andrew_Carr24
# ex_token is the api token returned from api_credentials_to_token.

get_user_network(screen_name = "Andrew_Carr24", degree = 2, token = ex_token)
}
}
